Use number of characters for unicode text
Before this commit, the length of the message was set to the length of
the input string. However, when this is encoded as unicode, the length
of the message should be the number of bytes in the raw input.

This is now fixed by making encodeData return the number of input
characters being encoded.